Change the name of custom callback, create professori module.
Former-commit-id: daa70c8270287c3b87b0d1b369788ab8b32b19ff
