+++ functions.py
@@ -23,6 +23,7 @@ import logging
 
 from module.lezioni import lezioni_cmd
 from module.esami import esami_cmd
+from module.professori import prof_cmd
 
 # Debug
 disable_chatid_logs = 1 #news, stats
@@ -129,46 +130,10 @@ def read_md(namefile):
     in_file.close()
     return text
 
-def custom_callback(bot, update, cmd):
+def informative_callback(bot, update, cmd):
     checkLog(bot, update, cmd)
     messageText = read_md(cmd)
     bot.sendMessage(chat_id=update.message.chat_id, text=messageText, parse_mode='Markdown')
-
-def prof_output(prof):
-    output = "Ruolo: " + prof["Ruolo"] + "\n"
-    output += "Cognome: " + prof["Cognome"] + "\n"
-    output += "Nome: " + prof["Nome"] + "\n"
-    output += "Indirizzo email : " + prof["Email"] + "\n"
-    output += "Sito web: " + prof["Sito"] + "\n"
-    output += "Scheda DMI: " + prof["SchedaDMI"] + "\n"
-    return output
-
-def prof_cmd(profs):
-
-    output_str = "Poffarbacco, qualcosa non va. Segnalalo ai dev \contributors \n"
-
-    if(profs):
-
-        output = Set()
-        profs = [x.lower().encode('utf-8') for x in profs if len(x) > 3]
-
-        with open("data/json/professori.json") as data_file:
-            professori_data = json.load(data_file)
-
-        for prof in profs:
-            professori = [professore for professore in professori_data if (prof in professore["Nome"].lower() or prof in professore["Cognome"].lower())]
-            for professore in professori:
-                output.add(prof_output(professore))
-
-        if(len(output)):
-            output_str = '\n'.join(list(output))
-        else:
-            output_str = "Nessun risultato trovato :(\n"
-
-    else:
-        output_str = "La sintassi del comando è: /prof <nomeprofessore>\n"
-
-    return output_str
 
 def rapp_cmd():
 	output = "Usa uno dei seguenti comandi per mostrare i rispettivi rappresentanti\n"
